requirements.txt
asyncio
websockets
datetime
torchvision
pillow
python-dotenv